Moved airlines app to tck
