+++ api/pom.xml
@@ -33,11 +33,6 @@
             <artifactId>org.osgi.annotation.versioning</artifactId>
             <version>1.0.0</version>
         </dependency>
-        <dependency>
-            <groupId>javax.ws.rs</groupId>
-            <artifactId>javax.ws.rs-api</artifactId>
-            <version>2.0</version>
-        </dependency>
     </dependencies>
 
     <build>
