+++ tck/pom.xml
@@ -99,5 +99,11 @@
             <artifactId>arquillian-junit-container</artifactId>
             <scope>compile</scope>
         </dependency>
+
+        <dependency>
+            <groupId>javax.ws.rs</groupId>
+            <artifactId>javax.ws.rs-api</artifactId>
+            <version>2.0</version>
+        </dependency>
     </dependencies>
 </project>
