+++ tck/src/main/java/org/eclipse/microprofile/openapi/apps/airlines/exception/NotFoundException.java
@@ -14,11 +14,11 @@
  *  limitations under the License.
  */
 
-package org.eclipse.microprofile.openapi.samples.jaxrs.exception;
+package org.eclipse.microprofile.openapi.apps.airlines.exception;
 
-public class BadRequestException extends ApiException{
+public class NotFoundException extends ApiException {
   private int code;
-  public BadRequestException (int code, String msg) {
+  public NotFoundException (int code, String msg) {
     super(code, msg);
     this.code = code;
   }
