+++ tck/src/main/java/org/eclipse/microprofile/openapi/apps/airlines/model/CreditCard.java
@@ -13,9 +13,10 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-package org.eclipse.microprofile.openapi.samples.jaxrs.model;
+package org.eclipse.microprofile.openapi.apps.airlines.model;
 
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
+
 public class CreditCard {
 
     @Schema(required = true, example="VISA")
