+++ tck/src/main/java/org/eclipse/microprofile/openapi/apps/airlines/resources/package-info.java
@@ -12,4 +12,4 @@
  */
 
 @org.osgi.annotation.versioning.Version("1.0")
-package org.eclipse.microprofile.openapi.samples.jaxrs.model;
+package org.eclipse.microprofile.openapi.apps.airlines.resources;
